Refactors code to have the WebSocket handler in a separated class and file.
